assistant: remove the Agent Wizard skill
Its instructions were outdated. The wizard-only tools (createAgent, getAgentRules)
go with it; SyncAutomation moves to the agent skill's operations, where the
AgentProperties subscriptions toggle still needs it.
