--- packages/core/compute/assistant-toolkit/src/skills/agent-wizard/index.ts
@@ -1,6 +1,0 @@
-//
-// Copyright 2025 DXOS.org
-//
-
-export { default as AgentWizardSkill } from './skill';
-export { AgentWizardHandlers, AgentWizardOperations } from './operations';
