--- packages/core/compute/assistant-toolkit/src/skills/agent-wizard/operations/index.ts
@@ -1,16 +1,0 @@
-//
-// Copyright 2026 DXOS.org
-//
-
-import * as Operation from '@dxos/compute/Operation';
-import * as OperationHandlerSet from '@dxos/compute/OperationHandlerSet';
-
-import { AgentRules, CreateAgent, SyncAutomation } from './definitions';
-
-export * as AgentWizardOperations from './definitions';
-
-export const AgentWizardHandlers = OperationHandlerSet.lazy([
-  CreateAgent.pipe(Operation.lazyHandler(() => import('./create-agent'))),
-  AgentRules.pipe(Operation.lazyHandler(() => import('./agent-rules'))),
-  SyncAutomation.pipe(Operation.lazyHandler(() => import('./sync-automation'))),
-]);
