+++ packages/core/compute/assistant-toolkit/src/skills/agent/operations/index.ts
@@ -5,11 +5,12 @@
 import * as Operation from '@dxos/compute/Operation';
 import * as OperationHandlerSet from '@dxos/compute/OperationHandlerSet';
 
-import { GetContext, Relay } from './definitions';
+import { GetContext, Relay, SyncAutomation } from './definitions';
 
 export * as AgentSkillOperations from './definitions';
 
 export const AgentSkillHandlers = OperationHandlerSet.lazy([
   GetContext.pipe(Operation.lazyHandler(() => import('./get-context'))),
   Relay.pipe(Operation.lazyHandler(() => import('./relay'))),
+  SyncAutomation.pipe(Operation.lazyHandler(() => import('./sync-automation'))),
 ]);
