+++ packages/core/compute/assistant-toolkit/src/skills/index.ts
@@ -3,7 +3,6 @@
 //
 
 export * from './agent';
-export * from './agent-wizard';
 export * from './alarm';
 export * from './automation';
 export * from './skill-manager';
