+++ packages/core/compute/assistant-toolkit/src/testing/operations.ts
@@ -15,7 +15,6 @@ import { Text } from '@dxos/schema';
 import { Employer, Organization, Outline, Person, Task, TaskSet } from '@dxos/types';
 
 import { AgentHandlers } from '../operations';
-import { AgentWizardHandlers } from '../skills/agent-wizard/operations';
 import { AgentSkillHandlers } from '../skills/agent/operations';
 import { AlarmHandlers } from '../skills/alarm/operations';
 import { ChatContextHandlers } from '../skills/chat-context/operations';
@@ -35,7 +34,6 @@ export const OperationTestLayer = AssistantTestLayerWithTriggers({
   operationHandlers: OperationHandlerSet.merge(
     AgentHandlers,
     AgentSkillHandlers,
-    AgentWizardHandlers,
     AlarmHandlers,
     ChatContextHandlers,
     DelegationSkillHandlers,
