+++ packages/plugins/plugin-assistant/src/containers/AgentProperties/AgentProperties.tsx
@@ -6,7 +6,7 @@ import React, { useCallback, useRef } from 'react';
 
 import { useSpaceCallback } from '@dxos/app-framework/ui';
 import { type AppSurface } from '@dxos/app-toolkit/ui';
-import { Agent, AgentWizardOperations } from '@dxos/assistant-toolkit';
+import { Agent, AgentSkillOperations } from '@dxos/assistant-toolkit';
 import * as Operation from '@dxos/compute/Operation';
 import { Obj, Ref } from '@dxos/echo';
 import { log } from '@dxos/log';
@@ -26,7 +26,7 @@ export const AgentProperties = ({ subject: agent }: AgentPropertiesProps) => {
     spaceId,
     [] as const,
     () =>
-      Operation.invoke(AgentWizardOperations.SyncAutomation, {
+      Operation.invoke(AgentSkillOperations.SyncAutomation, {
         agent: Ref.make(agent),
         subscriptions: pendingSubscriptions.current,
       }),
